Affinity: Update resource lifetime
Update resource lifetime discussion in the Affinity proposal, based on https://github.com/codeplaysoftware/standards-proposals/issues/67 discussion.

NOTE: I have not changed the rest of the proposal to make it consistent with these changes. If people like this pull request, I will then update the rest of the proposal accordingly.